+++ CHANGELOG.markdown
@@ -1,9 +1,9 @@
 CHANGELOG for the Transifex API Package
 ===============
 
-* 1.3.0 (2015-XX-XX)
+* 1.3.0 (2016-XX-XX)
 
- * Undeprecated `TransifexObject::processResponse()`
+ * Deprecated `Http` class and `joomla/http` package integration
 
 * 1.2.0 (2015-07-13)